✨ Added `--skip` flag to exclude post slugs from being fetched
A post fragment is sometimes still too big and can't be broken down further. Algolia throws an error in that case and stops saving the rest of the fragments. In order to the rest of the fragments saved, it's easiest to exclude the post slug from being fetched in first place. It can be updated and indexed using the webhooks later on.